refact: renamed `objectContains` to `isPartialOf`, add documentation for him and add more tests
